+++ lua/onedark/types/init.lua
@@ -1,0 +1,5 @@
+local types = {}
+
+types.od = require('onedark.types.od')
+
+return types