+++ lua/onedark/types/od.lua
@@ -1,0 +1,9 @@
+---@diagnostic disable: duplicate-doc-class
+local od = {}
+
+---@alias od.TerminalConf "'conf'" | "'yml'"
+od.TerminalConf = {}
+od.TerminalConf.kitty = 'conf'
+od.TerminalConf.alacritty = 'yml'
+
+return od